remove debug code, change OSX building

+++ recipes/bioconductor-iclusterplus/build.sh
@@ -2,7 +2,9 @@
 mv DESCRIPTION DESCRIPTION.old
 grep -v '^Priority: ' DESCRIPTION.old > DESCRIPTION
 mkdir -p ~/.R
-ln -s $FC ${PREFIX}/bin/gfortran
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    ln -s $FC ${PREFIX}/bin/gfortran
+fi
 echo -e "CC=$CC
 FC=$FC
 CXX=$CXX
@@ -12,4 +14,6 @@ CXX14=$CXX" > ~/.R/Makevars
 sed -i.bak "1652i\\
   FC=\"gfortran\"" configure
 $R CMD INSTALL --build .
-rm ${PREFIX}/bin/gfortran
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    rm ${PREFIX}/bin/gfortran
+fi
